+++ network/base/views.py
@@ -1,5 +1,6 @@
 import ephem
 from datetime import datetime, timedelta
+from StringIO import StringIO
 
 from django.conf import settings
 from django.contrib import messages
@@ -7,18 +8,23 @@ from django.views.decorators.http import require_POST
 from django.shortcuts import get_object_or_404, render, redirect
 from django.core.urlresolvers import reverse
 from django.utils.timezone import now, make_aware, utc
-from django.http import JsonResponse
+from django.http import JsonResponse, HttpResponseNotFound, HttpResponseServerError, HttpResponse
 from django.contrib.auth.decorators import login_required
+from django.core.management import call_command
 
-from network.base.models import (Station, Transponder, Observation,
+from network.base.models import (Station, Transmitter, Observation,
                                  Data, Satellite, Antenna)
 from network.base.forms import StationForm
+from network.base.decorators import admin_required
 
 
 def index(request):
     """View to render index page."""
     observations = Observation.objects.all()
-    featured_station = Station.objects.filter(active=True).latest('featured_date')
+    try:
+        featured_station = Station.objects.filter(active=True).latest('featured_date')
+    except Station.DoesNotExist:
+        featured_station = None
 
     ctx = {
         'latest_observations': observations.filter(end__lt=now()),
@@ -29,6 +35,43 @@ def index(request):
     }
 
     return render(request, 'base/home.html', ctx)
+
+
+def custom_404(request):
+    """Custom 404 error handler."""
+    return HttpResponseNotFound(render(request, '404.html'))
+
+
+def custom_500(request):
+    """Custom 500 error handler."""
+    return HttpResponseServerError(render(request, '500.html'))
+
+
+def robots(request):
+    data = render(request, 'robots.txt', {'environment': settings.ENVIRONMENT})
+    response = HttpResponse(data,
+                            content_type='text/plain; charset=utf-8')
+    return response
+
+
+@admin_required
+def settings_site(request):
+    """View to render settings page."""
+    if request.method == 'POST':
+        if request.POST['fetch']:
+            try:
+                out = StringIO()
+                call_command('fetch_data', stdout=out)
+                request.session['fetch_out'] = out.getvalue()
+            except:
+                messages.error(request, 'fetch command failed.')
+        return redirect(reverse('base:settings_site'))
+
+    fetch_out = request.session.get('fetch_out', False)
+    if fetch_out:
+        del request.session['fetch_out']
+        return render(request, 'base/settings_site.html', {'fetch_data': fetch_out})
+    return render(request, 'base/settings_site.html')
 
 
 def observations_list(request):
@@ -44,14 +87,14 @@ def observation_new(request):
     me = request.user
     if request.method == 'POST':
         sat_id = request.POST.get('satellite')
-        trans_id = request.POST.get('transponder')
+        trans_id = request.POST.get('transmitter')
         start_time = datetime.strptime(request.POST.get('start-time'), '%Y-%m-%d %H:%M')
         start = make_aware(start_time, utc)
         end_time = datetime.strptime(request.POST.get('end-time'), '%Y-%m-%d %H:%M')
         end = make_aware(end_time, utc)
         sat = Satellite.objects.get(norad_cat_id=sat_id)
-        trans = Transponder.objects.get(id=trans_id)
-        obs = Observation(satellite=sat, transponder=trans,
+        trans = Transmitter.objects.get(id=trans_id)
+        obs = Observation(satellite=sat, transmitter=trans,
                           author=me, start=start, end=end)
         obs.save()
 
@@ -71,19 +114,19 @@ def observation_new(request):
 
         return redirect(reverse('base:observation_view', kwargs={'id': obs.id}))
 
-    satellites = Satellite.objects.filter(transponder__alive=True)
-    transponders = Transponder.objects.filter(alive=True)
+    satellites = Satellite.objects.filter(transmitters__alive=True).distinct()
+    transmitters = Transmitter.objects.filter(alive=True)
 
     return render(request, 'base/observation_new.html',
                   {'satellites': satellites,
-                   'transponders': transponders,
+                   'transmitters': transmitters,
                    'date_min_start': settings.DATE_MIN_START,
                    'date_max_range': settings.DATE_MAX_RANGE})
 
 
 def prediction_windows(request, sat_id, start_date, end_date):
     try:
-        sat = Satellite.objects.filter(transponder__alive=True).filter(norad_cat_id=sat_id).get()
+        sat = Satellite.objects.filter(transmitters__alive=True).distinct().get(norad_cat_id=sat_id)
     except:
         data = {
             'error': 'You should select a Satellite first.'