+++ Changelog.md
@@ -1,3 +1,21 @@
+## 0.4
+
+* Rename Transponders to Transmitters
+* Hide unused API endpoints
+
+## 0.3
+
+* Last Seen feature on Ground Stations
+* Enhance Observations UI
+* Opbeat
+* Add custom 404/500 pages
+* robots.txt
+* Add site settings for fetching data
+
+## 0.2
+
+* UI improvements
+
 ## 0.1
 
 * Initial release